refactor: move types to submit module
